+++ backend/app/main.py
@@ -1,4 +1,5 @@
 # app run here
+import logging
 from fastapi import FastAPI
 from contextlib import asynccontextmanager
 
@@ -8,33 +9,27 @@ from backend.app.api.routes import router
 from backend.app.background_service import scheduler
 
 
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
 ws_manager: WSManager = WSManager()
 
 
 @asynccontextmanager
 async def lifespan(fastapi_app: FastAPI):
-    # Lấy dữ liệu từ db và xay dưng instance của WSManager
-    print("Start...")
+    """
+    Lấy dữ liệu từ db và xay dưng instance của WSManager
+    """
+    logging.info("Start...")
     scheduler.start()
     await ws_manager.add_information()
     
     yield
-    print("Close...")
+    logging.info("Close...")
     scheduler.shutdown()
 
 
 ws_manager: WSManager = WSManager()
-
-
-@asynccontextmanager
-async def lifespan(fastapi_app: FastAPI):
-    # Lấy dữ liệu từ db và xau dưng instance của WSManager
-    print("Start...")
-    await ws_manager.add_information()
-
-    yield
-    print("Close...")
-
 
 app = FastAPI(
     title=settings.PROJECT_NAME,
